Implemented AI player by ChatGPT
